add rename test files
